+++ python/packages/brilliant_ble/src/brilliant_ble/__init__.py
@@ -5,5 +5,6 @@ __all__ = ["brilliant_ble"]
 
 from .brilliant_ble import BrilliantBle
 from .brilliant_ble import BrilliantDeviceType
+from .brilliant_ble import OtaError
 
-__version__ = "3.0.0"
+__version__ = "3.1.0"